+++ engine/launcher/args/args.go
@@ -26,6 +26,11 @@ type EngineServerArgs struct {
 
 	//User consent to send metrics
 	DidUserAcceptSendingMetrics bool `json:"didUserAcceptSendingMetrics"`
+
+	KurtosisBackendType KurtosisBackendType `json:"kurtosisBackendType"`
+
+	// Should be deserialized differently depending on value of KurtosisBackendType
+	KurtosisBackendConfig interface{} `json:"kurtosisBackendConfig"`
 }
 
 // Even though the fields are public due to JSON de/serialization requirements, we still have this constructor so that
@@ -37,16 +42,19 @@ func NewEngineServerArgs(
 	imageVersionTag string,
 	metricsUserID string,
 	didUserAcceptSendingMetrics bool,
+	kurtosisBackendType KurtosisBackendType,
+	kurtosisBackendConfig interface{},
 ) (*EngineServerArgs, error) {
 	result := &EngineServerArgs{
-		GrpcListenPortNum:              grpcListenPortNum,
-		GrpcProxyListenPortNum:         grpcProxyListenPortNum,
-		LogLevelStr:                    logLevelStr,
-		ImageVersionTag:                imageVersionTag,
-		MetricsUserID:                  metricsUserID,
-		DidUserAcceptSendingMetrics:    didUserAcceptSendingMetrics,
+		GrpcListenPortNum:           grpcListenPortNum,
+		GrpcProxyListenPortNum:      grpcProxyListenPortNum,
+		LogLevelStr:                 logLevelStr,
+		ImageVersionTag:             imageVersionTag,
+		MetricsUserID:               metricsUserID,
+		DidUserAcceptSendingMetrics: didUserAcceptSendingMetrics,
+		KurtosisBackendType:         kurtosisBackendType,
+		KurtosisBackendConfig:       kurtosisBackendConfig,
 	}
-
 	if err := result.validate(); err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred validating engine server args")
 	}