+++ engine/server/engine/main.go
@@ -7,9 +7,13 @@ package main
 
 import (
 	"fmt"
+	"github.com/kurtosis-tech/container-engine-lib/lib"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/backend_creator"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface"
+	"github.com/kurtosis-tech/kurtosis-core/launcher/api_container_launcher"
 	"github.com/kurtosis-tech/kurtosis-engine-server/api/golang/kurtosis_engine_rpc_api_bindings"
 	"github.com/kurtosis-tech/kurtosis-engine-server/launcher/args"
+	"github.com/kurtosis-tech/kurtosis-engine-server/launcher/args/kurtosis_backend_config"
 	"github.com/kurtosis-tech/kurtosis-engine-server/server/engine/enclave_manager"
 	"github.com/kurtosis-tech/kurtosis-engine-server/server/engine/server"
 	metrics_client "github.com/kurtosis-tech/metrics-library/golang/lib/client"
@@ -29,16 +33,15 @@ const (
 	grpcServerStopGracePeriod = 5 * time.Second
 
 	shouldFlushMetricsClientQueueOnEachEvent = false
-
 )
 
 // Nil indicates that the KurtosisBackend should not operate in API container mode, which is appropriate here
 //  because this isn't the API container
 var apiContainerModeArgsForKurtosisBackend *backend_creator.APIContainerModeArgs = nil
 
-type doNothingMetricsClientCallback struct {}
+type doNothingMetricsClientCallback struct{}
 
-func (d doNothingMetricsClientCallback) Success() {}
+func (d doNothingMetricsClientCallback) Success()          {}
 func (d doNothingMetricsClientCallback) Failure(err error) {}
 
 func main() {
@@ -57,7 +60,7 @@ func main() {
 	os.Exit(successExitCode)
 }
 
-func runMain () error {
+func runMain() error {
 	serverArgs, err := args.GetArgsFromEnv()
 	if err != nil {
 		return stacktrace.Propagate(err, "Couldn't retrieve engine server args from the environment")
@@ -69,12 +72,15 @@ func runMain () error {
 	}
 	logrus.SetLevel(logLevel)
 
-	kurtosisBackend, err := backend_creator.GetLocalDockerKurtosisBackend(apiContainerModeArgsForKurtosisBackend)
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred getting a Kurtosis backend connected to local Docker")
+	backendConfig := serverArgs.KurtosisBackendConfig
+	if backendConfig == nil {
+		return stacktrace.NewError("Backend configuration parameters are null - there must be backend configuration parameters.")
 	}
 
-	enclaveManager := enclave_manager.NewEnclaveManager(kurtosisBackend)
+	enclaveManager, err := getEnclaveManager(serverArgs.KurtosisBackendType, backendConfig)
+	if err != nil {
+		return stacktrace.Propagate(err, "Failed to create an enclave manager for backend type '%v' and config '%+v'", serverArgs.KurtosisBackendType, backendConfig)
+	}
 
 	metricsClient, metricsClientCloseFunc, err := metrics_client.CreateMetricsClient(
 		source.KurtosisEngineSource,
@@ -111,4 +117,35 @@ func runMain () error {
 		return stacktrace.Propagate(err, "An error occurred running the server.")
 	}
 	return nil
+}
+
+func getEnclaveManager(kurtosisBackendType args.KurtosisBackendType, backendConfig interface{}) (*enclave_manager.EnclaveManager, error){
+	var kurtosisBackend backend_interface.KurtosisBackend
+	var err error
+	var apiContainerKurtosisBackendConfigSupplier api_container_launcher.KurtosisBackendConfigSupplier
+	switch kurtosisBackendType {
+	case args.KurtosisBackendType_Docker:
+		kurtosisBackend, err = backend_creator.GetLocalDockerKurtosisBackend(apiContainerModeArgsForKurtosisBackend)
+		if err != nil {
+			return nil, stacktrace.Propagate(err, "An error occurred getting local Docker Kurtosis backend")
+		}
+		apiContainerKurtosisBackendConfigSupplier = api_container_launcher.NewDockerKurtosisBackendConfigSupplier()
+	case args.KurtosisBackendType_Kubernetes:
+		kubernetesBackendConfig, ok := (backendConfig).(kurtosis_backend_config.KubernetesBackendConfig)
+		if !ok {
+			return nil, stacktrace.NewError("Failed to cast cluster configuration interface to the appropriate type, even though Kurtosis backend type is '%v'", args.KurtosisBackendType_Kubernetes.String())
+		}
+		// TODO TODO TODO Change enclave size in GB to take a k8s friendly string to allow flexible size specs
+		kurtosisBackend, err = lib.GetLocalKubernetesKurtosisBackend(kubernetesBackendConfig.StorageClass, int(kubernetesBackendConfig.EnclaveSizeInGigabytes))
+		if err != nil {
+			return nil, stacktrace.Propagate(err, "Failed to get a Kubernetes backend with storage class '%v' and enclave size (in GB) %d", kubernetesBackendConfig.StorageClass, kubernetesBackendConfig.EnclaveSizeInGigabytes)
+		}
+		apiContainerKurtosisBackendConfigSupplier = api_container_launcher.NewKubernetesKurtosisBackendConfigSupplier(kubernetesBackendConfig.StorageClass, kubernetesBackendConfig.EnclaveSizeInGigabytes)
+	default:
+		return nil, stacktrace.NewError("Backend type '%v' was not recognized by engine server.", kurtosisBackendType.String())
+	}
+
+	enclaveManager := enclave_manager.NewEnclaveManager(kurtosisBackend, apiContainerKurtosisBackendConfigSupplier)
+
+	return enclaveManager, nil;
 }